chore(configs): separate configs from bin
