+++ packages/configs/src/utils/printInfo.js
